Create the project structure
Created the index page of the website, a README file and the base of the
project. This consist in two folders, one for the images and the other
for the pages, called "recipes".

Also created in the recipes folder a webpage for each item that will be
linked in the index page.