Add images, recipes structure and carbonara content
Added for each webpage an image under the title showing the dish in the
recipe, all the images have width and height value = 300.

Created for each recipe page 3 sections: description, ingredients and
steps. These sections titles are heading 3.

Added the content of carbonara page: a two paragraph description, an
unordered list for the ingredients and an ordered list for the steps.