Change style and format code
Changed the style of the 4 webpages. This project does not use CSS, so
I use the tags <center> and <tt> even if HTML5 does not support them.

The source code of the 4 pages was reviewed to check and correct
indentation, and to give uniformity in separating the various elements
of the code itself.